Returns non-message templates as changeTopic templates
