+++ lib/helpers/broadcast.js
@@ -62,22 +62,6 @@ function getById(broadcastId, resetCache = false) {
 
 /**
  * @param {Object} contentfulEntry
- * @return {Object}
- */
-function parseBroadcastInfoFromContentfulEntry(contentfulEntry) {
-  const result = {
-    id: contentful.getContentfulIdFromContentfulEntry(contentfulEntry),
-    name: contentful.getNameTextFromContentfulEntry(contentfulEntry),
-    type: contentful.getContentTypeFromContentfulEntry(contentfulEntry),
-    createdAt: contentfulEntry.sys.createdAt,
-    updatedAt: contentfulEntry.sys.updatedAt,
-  };
-  logger.debug('parseBroadcastInfoFromContentfulEntry', { result });
-  return result;
-}
-
-/**
- * @param {Object} contentfulEntry
  * @return {Promise}
  */
 function parseBroadcastFromContentfulEntry(contentfulEntry) {
@@ -90,7 +74,7 @@ function parseBroadcastFromContentfulEntry(contentfulEntry) {
     return module.exports.parseLegacyBroadcastFromContentfulEntry(contentfulEntry);
   }
 
-  const data = module.exports.parseBroadcastInfoFromContentfulEntry(contentfulEntry);
+  const data = helpers.contentfulEntry.getNameInfoFromContentfulEntry(contentfulEntry);
   const message = module.exports
     .parseBroadcastMessageFromContentfulEntryAndTemplateName(contentfulEntry, contentType);
   const fieldNames = contentTypeConfigs[contentType].templates;
@@ -108,7 +92,7 @@ function parseBroadcastFromContentfulEntry(contentfulEntry) {
  * @return {Promise}
  */
 function parseLegacyBroadcastFromContentfulEntry(contentfulEntry) {
-  const data = module.exports.parseBroadcastInfoFromContentfulEntry(contentfulEntry);
+  const data = helpers.contentfulEntry.getNameInfoFromContentfulEntry(contentfulEntry);
   data.message = {
     text: contentfulEntry.fields.message,
     attachments: contentful.getAttachmentsFromContentfulEntry(contentfulEntry),
@@ -176,7 +160,6 @@ module.exports = {
   getById,
   getContentTypes,
   parseBroadcastFromContentfulEntry,
-  parseBroadcastInfoFromContentfulEntry,
   parseBroadcastMessageFromContentfulEntryAndTemplateName,
   parseLegacyBroadcastFromContentfulEntry,
   parseTemplatesFromContentfulEntryAndFieldNames,
