+++ lib/helpers/contentfulEntry.js
@@ -1,5 +1,6 @@
 'use strict';
 
+const logger = require('winston');
 const contentful = require('../contentful');
 const helpers = require('../helpers');
 
@@ -38,6 +39,22 @@ function parseContentfulEntry(contentfulEntry) {
 
 /**
  * @param {Object} contentfulEntry
+ * @return {Object}
+ */
+function getNameInfoFromContentfulEntry(contentfulEntry) {
+  const result = {
+    id: contentful.getContentfulIdFromContentfulEntry(contentfulEntry),
+    name: contentful.getNameTextFromContentfulEntry(contentfulEntry),
+    type: contentful.getContentTypeFromContentfulEntry(contentfulEntry),
+    createdAt: contentfulEntry.sys.createdAt,
+    updatedAt: contentfulEntry.sys.updatedAt,
+  };
+  logger.debug('parseNameInfoFromContentfulEntry', { result });
+  return result;
+}
+
+/**
+ * @param {Object} contentfulEntry
  * @param {String} templateName
  * @return {Object}
  */
@@ -46,7 +63,7 @@ function getMessageTemplateFromContentfulEntryAndTemplateName(contentfulEntry, t
   if (contentType !== 'message') {
     return {
       template: 'changeTopic',
-      topic: contentful.getContentTypeFromContentfulEntry(contentfulEntry),
+      topic: module.exports.getNameInfoFromContentfulEntry(contentfulEntry),
     };
   }
   return {
@@ -59,5 +76,6 @@ function getMessageTemplateFromContentfulEntryAndTemplateName(contentfulEntry, t
 module.exports = {
   getById,
   getMessageTemplateFromContentfulEntryAndTemplateName,
+  getNameInfoFromContentfulEntry,
   parseContentfulEntry,
 };
